+++ SubEthaEdit-Mac/DNDArrayController.m
@@ -111,7 +111,7 @@ NSString *MovedRowsType = @"MOVED_ROWS_TYPE";
 {
 	
     NSArray		*objects = [self arrangedObjects];
-	int			index = [indexSet lastIndex];
+	NSUInteger	   index = [indexSet lastIndex];
 	
     int			aboveInsertIndexCount = 0;
     id			object;
